// frontend/src/type.ts
export type Card = {
  abilities: string
  cost: number | null
  createdAt: string
  draw: number | null
  faction: string
  gold: number | null
  health: number | null
  id: number
  image: any
  initiative: number | null
  melee_attack: number | null
  name: string
  prestige: number | null
  publishedAt: string
  ranged_attack: number | null
  riposte: number | null
  subheading: string | null
  subtype: string | null
  trait: string | null
  type: string
  updatedAt: string
}